requirements.txt
flask==3.1.2
flask-SocketIO==5.5.1
requests==2.32.5
psutil==7.2.2
humanize==4.15.0
pyinstaller==6.19.0
pillow==12.1.1
scikit-learn==1.8.0